move setup function back into main.go
